+++ client/src/components/PlanDetail/PlanElement/AccountBook/AccountItem.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import Icon from '../../../Common/Icon'
+import Icon from '../../../Commons/Icon'
 import styles from './AccountItem.module.scss'
 
 type AccountItemProps = {
