+++ client/src/components/PlanDetail/PlanElement/Tapbar.tsx
@@ -1,5 +1,5 @@
 import styles from './Tapbar.module.scss'
-import Icon from '../../Common/Icon'
+import Icon from '../../Commons/Icon'
 
 type TapbarViewProps = {
   currentTap: string
